CLIM-1454: Atom20 Split prepare-raster into lib/map/image-rastering namespace
Pure structural move, no behaviour change. Splits apps/src/lib/prepare-raster.ts
(sender half, receiver half, and a debug class that were never live in the same
page load) into apps/src/lib/map/image-rastering/, one primary export per file,
mirroring lib/download/'s verb-first-filename + types.ts + barrel precedent.

prepareRaster's tail is extracted into a named signalRasterReady, called
unconditionally (the SignalReady parameter lands in Atom 23). The debugger
class gains a createRasterDebugger() factory wrapper for the barrel and a
future lazy import() (Atom 25); the class itself stays exported and is still
constructed eagerly by download-map-modal.tsx this Atom.

One substantive fix carried along: waitForMapsSettled and waitForMarkerIcons
both carried a wrong '@remark Where does this run?: In the user's browser.'
Both only ever run from prepareRaster, which executes in the screenshot
service's headless browser -- corrected to say so, since the whole split is
organized around which browser runs which half.
